+++ framework/modules/saf_sofa_reader/libmysofa/config.h
@@ -1,0 +1,10 @@
+
+#if !defined _CONFIG_H
+#define _CONFIG_H
+
+#define CMAKE_INSTALL_PREFIX "${CMAKE_INSTALL_PREFIX}"
+#define CPACK_PACKAGE_VERSION_MAJOR ${CPACK_PACKAGE_VERSION_MAJOR}
+#define CPACK_PACKAGE_VERSION_MINOR ${CPACK_PACKAGE_VERSION_MINOR}
+#define CPACK_PACKAGE_VERSION_PATCH ${CPACK_PACKAGE_VERSION_PATCH}
+
+#endif
