+++ framework/modules/saf_sofa_reader/libmysofa/hdf/reader.h
@@ -1,46 +1,13 @@
 /*
- Copyright (c) 2016, Symonics GmbH, Christian Hoene
- All rights reserved.
 
- Redistribution and use in source and binary forms, with or without
- modification, are permitted provided that the following conditions are
- met:
+ Copyright 2016 Christian Hoene, Symonics GmbH
 
-     (1) Redistributions of source code must retain the above copyright
-     notice, this list of conditions and the following disclaimer.
+ */
 
-     (2) Redistributions in binary form must reproduce the above copyright
-     notice, this list of conditions and the following disclaimer in
-     the documentation and/or other materials provided with the
-     distribution.
+#ifndef READER_H_
+#define READER_H_
 
-     (3)The name of the author may not be used to
-     endorse or promote products derived from this software without
-     specific prior written permission.
-
- THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR
- IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- DISCLAIMED. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT,
- INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
- IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- POSSIBILITY OF SUCH DAMAGE.
-*/
-
-#ifndef MYSOFA_HDF_READER_H_
-#define MYSOFA_HDF_READER_H_
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#if defined(SAF_ENABLE_SOFA_READER_MODULE)
-
-#include "../mysofa.h"
+#include "../hrtf/mysofa.h"
 #include <stdint.h>
 #include <stdio.h>
 
@@ -203,7 +170,11 @@ void gcolFree(struct GCOL *gcol);
 int treeRead(struct READER *reader, struct DATAOBJECT *data);
 
 struct READER {
-  FILE *fhd;
+  FILE *fhd; //< file handler if file is used
+
+  const char *memory; //< memory buffer if memory shall be used
+  uint64_t memory_pos;
+  uint64_t memory_len;
 
   struct DATAOBJECT *all;
 
@@ -219,10 +190,10 @@ uint64_t readValue(struct READER *reader, int size);
 
 int gunzip(int inlen, char *in, int *outlen, char *out);
 
-#endif /* SAF_ENABLE_SOFA_READER_MODULE */
+char *mysofa_strdup(const char *s);
+int mysofa_read(struct READER *reader, void *buf, size_t n);
+int mysofa_seek(struct READER *reader, long offset, int whence);
+long mysofa_tell(struct READER *reader);
+int mysofa_getc(struct READER *reader);
 
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* MYSOFA_HDF_READER_H_ */
+#endif /* READER_H_ */
