+++ framework/modules/saf_sofa_reader/libmysofa/hrtf/mysofa_export.h
@@ -1,0 +1,6 @@
+
+/**
+ * use this file if you do not use CMAKE to generate you an own file
+ */
+
+#define MYSOFA_EXPORT
